requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
atlassian-python-api==3.41.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
Deprecated==1.2.14
dnspython==2.4.2
Flask==2.3.3
frozenlist==1.4.0
fsspec==2023.9.0
html2text==2020.1.16
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
langchain==0.0.285
langsmith==0.0.35
llama-hub==0.0.29
llama-index==0.8.23.post1
loguru==0.7.1
MarkupSafe==2.1.3
marshmallow==3.20.1
mongoengine==0.27.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
nltk==3.8.1
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.2
openai==0.28.0
packaging==23.1
pandas==2.1.0
pinecone-client==2.2.2
plotly==5.16.1
psutil==5.9.5
pydantic==2.3.0
pydantic_core==2.6.3
pymongo==4.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.20
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.7
wrapt==1.15.0
yarl==1.9.2
